+++ src/test/java/org/osgl/issues/Gh78.java
@@ -1,0 +1,41 @@
+package org.osgl.issues;
+
+/*-
+ * #%L
+ * Java Tool
+ * %%
+ * Copyright (C) 2014 - 2018 OSGL (Open Source General Library)
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+import org.junit.Test;
+import org.osgl.$;
+import org.osgl.TestBase;
+import org.osgl.util.C;
+
+import java.util.Map;
+
+public class Gh78 extends TestBase {
+
+    @Test
+    public void test() {
+        Map<String, String> from = C.Map("name", "Tom");
+        Map<String, String> to = C.newMap();
+        $.map(from).to(to);
+        eq(1, to.size());
+        eq("Tom", to.get("name"));
+    }
+
+}