+++ src/test/java/org/osgl/issues/g79/Bar.java
@@ -1,0 +1,8 @@
+package org.osgl.issues.g79;
+
+import java.util.List;
+import java.util.Map;
+
+public class Bar {
+    public Map<String, List<Foo>> fooMap;
+}