+++ src/test/java/org/osgl/issues/g79/Bean.java
@@ -1,0 +1,8 @@
+package org.osgl.issues.g79;
+
+import java.util.List;
+import java.util.Map;
+
+public class Bean {
+    public Map<String, List<Bar>> barMap;
+}