+++ src/test/java/org/osgl/issues/g79/Foo.java
@@ -1,0 +1,7 @@
+package org.osgl.issues.g79;
+
+import org.osgl.util.N;
+
+public class Foo {
+    public int id = N.randInt();
+}